requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
executing==1.0.0
flake8==5.0.4
httpie==3.2.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pygraphviz==1.10
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
shell==1.0.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5